move llama inference code to tensor_parallel
