+++ tests/test_infer/test_llama_infer.py
@@ -7,7 +7,7 @@ import numpy as np
 import colossalai
 from colossalai.logging import disable_existing_loggers
 from colossalai.testing import clear_cache_before_run, parameterize, rerun_if_address_is_in_use, spawn
-from llama_infer_engine import TPCacheManagerInferenceEngine
+from colossalai.tensor_parallel.tpinference.llama_infer_engine import TPCacheManagerInferenceEngine
 
 os.environ['TRANSFORMERS_NO_ADVISORY_WARNINGS'] = 'true'
 TPSIZE = 1
